Split AlongStepNeutral details into one per class
