+++ src/celeritas/global/alongstep/detail/NoELoss.hh
@@ -1,18 +1,14 @@
 //----------------------------------*-C++-*----------------------------------//
-// Copyright 2022-2023 UT-Battelle, LLC, and other Celeritas developers.
+// Copyright 2023 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/global/alongstep/detail/AlongStepNeutralImpl.hh
+//! \file celeritas/global/alongstep/detail/NoELoss.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include "corecel/Assert.hh"
 #include "corecel/Macros.hh"
-#include "corecel/Types.hh"
 #include "corecel/math/Quantity.hh"
-
-#include "../AlongStep.hh"
 
 namespace celeritas
 {
@@ -21,27 +17,6 @@ class CoreTrackView;
 
 namespace detail
 {
-//---------------------------------------------------------------------------//
-/*!
- * Helper class for along-step kernel when no MSC is in use.
- */
-struct NoMsc
-{
-    //! MSC never applies to the current track
-    CELER_FUNCTION bool
-    is_applicable(CoreTrackView const&, real_type step) const
-    {
-        CELER_ASSERT(step > 0);
-        return false;
-    }
-
-    //! No updates needed to the physical and geometric step lengths
-    CELER_FUNCTION void limit_step(CoreTrackView const&, StepLimit*) const {}
-
-    //! MSC is never applied
-    CELER_FUNCTION void apply_step(CoreTrackView const&, StepLimit*) const {}
-};
-
 //---------------------------------------------------------------------------//
 /*!
  * Class that returns zero energy loss.
